feat(fetch-table): merge FetchTable & FetchListTable
